Review fixes for SSL_get_all_signature_algorithm_names(), plus a test
